# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gosso - SSO Authentication Server
  description: |
    A production-ready Single Sign-On (SSO) authentication server built with Go.

    Supports:
    - Username/password login with MFA (TOTP + backup codes)
    - OAuth 2.0 Authorization Server (authorization code, refresh token, client credentials)
    - OpenID Connect Provider (ID token, discovery, JWKS, userinfo)
    - Social login (Google, GitHub, WeChat)
    - Admin account management
    - Email/phone verification and password reset
  version: 1.0.0
  contact:
    name: rushairer
    url: https://github.com/rushairer/gosso
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with username and password
      description: |
        Authenticates a user with username and password.
        If MFA is enabled, returns an MFA token instead of full credentials.
        If MFA is not enabled, returns access token, refresh token, and session ID.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Response_LoginSuccess"
                  - $ref: "#/components/schemas/Response_MFARequired"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new access token and refresh token pair.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: Revokes the current session and access token. Requires a valid JWT.
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: logged out
  /api/auth/session:
    get:
      tags: [Auth]
      summary: Get current session
      description: Returns the current session details. Requires a valid JWT.
      operationId: getSession
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/sessions:
    get:
      tags: [Auth]
      summary: List all sessions
      description: Returns all active sessions for the authenticated account. Requires a valid JWT.
      operationId: listSessions
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a session
      description: Revokes a specific session by ID. Cannot revoke the current session (use /logout instead).
      operationId: revokeSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID to revoke
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data:
                  revoked: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify MFA code
      description: Verifies a TOTP or backup code during the MFA login flow.
      operationId: mfaVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAVerifyRequest"
      responses:
        "200":
          description: MFA verified, returns tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_TokenPairWithSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/mfa/enroll:
    post:
      tags: [Auth]
      summary: Enroll TOTP MFA
      description: Starts TOTP enrollment for the authenticated account. Returns a secret and provisioning URI.
      operationId: mfaEnroll
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAEnrollRequest"
      responses:
        "200":
          description: TOTP enrollment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/auth/mfa/activate:
    post:
      tags: [Auth]
      summary: Activate TOTP MFA
      description: Activates TOTP after verifying a code during enrollment.
      operationId: mfaActivate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAActivateRequest"
      responses:
        "200":
          description: TOTP activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: TOTP activated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/mfa:
    delete:
      tags: [Auth]
      summary: Disable TOTP MFA
      description: Disables TOTP MFA for the authenticated account.
      operationId: mfaDisable
      security:
        - BearerAuth: []
      responses:
        "200":
          description: MFA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: MFA disabled
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/mfa/backup-codes:
    post:
      tags: [Auth]
      summary: Generate backup codes
      description: Generates a new set of backup codes for the authenticated account. Previous codes are invalidated.
      operationId: mfaGenerateBackupCodes
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Backup codes generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      backup_codes:
                        type: array
                        items:
                          type: string
                        example: ["a1b2c3d4", "e5f6g7h8"]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/social/{provider}:
    get:
      tags: [Auth]
      summary: Get social login URL
      description: Returns a redirect to the OAuth2 authorization URL of the specified provider (google, github, wechat).
      operationId: socialAuthURL
      parameters:
        - $ref: "#/components/parameters/SocialProvider"
        - name: state
          in: query
          schema:
            type: string
          description: OAuth2 state parameter for CSRF protection
      responses:
        "302":
          description: Redirect to provider authorization URL
        "400":
          $ref: "#/components/responses/BadRequest"
        "501":
          description: Social login not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/social/{provider}/callback:
    get:
      tags: [Auth]
      summary: Social login callback
      description: |
        Handles the OAuth2 callback from the social login provider.
        Validates the `state` parameter against the `oauth_state` cookie (CSRF protection),
        then exchanges the authorization code for tokens.
      operationId: socialCallback
      parameters:
        - $ref: "#/components/parameters/SocialProvider"
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: OAuth2 authorization code from the provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State parameter for CSRF validation (must match oauth_state cookie)
      responses:
        "200":
          description: Social login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_TokenPairWithSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/verify/send:
    post:
      tags: [Auth]
      summary: Send verification code
      description: Sends a verification code to the specified email or phone number. Requires authentication.
      operationId: sendVerification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendVerificationRequest"
      responses:
        "200":
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: verification code sent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/verify/confirm:
    post:
      tags: [Auth]
      summary: Confirm verification code
      description: Confirms a verification code and marks the credential as verified.
      operationId: confirmVerification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmVerificationRequest"
      responses:
        "200":
          description: Credential verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: credential verified
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request password reset
      description: |
        Sends a password reset email if the email exists. Always returns 200 to prevent email enumeration.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent (or email not found — same response)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: if the email exists, a reset link has been sent
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Resets the password using a valid reset token.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: password has been reset successfully
        "400":
          $ref: "#/components/responses/BadRequest"

  # ──────────────────────────────────────────────
  # OAuth2 Clients
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login, token, or social login endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth with client_id as username and client_secret as password
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection via double-submit cookie pattern.
        1. GET any endpoint to receive `csrf_token` cookie
        2. Read cookie value and send as `X-CSRF-Token` header on POST/PUT/DELETE
        3. Not required when using `Authorization: Bearer` header (JWT auth)

  # ── Parameters ──
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: johndoe
        password:
          type: string
          format: password
          example: SecurePassword123!
...
    Response_LoginSuccess:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              description: Access token lifetime in seconds
              example: 900
            session_id:
              type: string
              format: uuid
...